macos fix test 1

+++ src/video_core/amdgpu/pm4_cmds.h
@@ -276,13 +276,15 @@ struct PM4CmdStrmoutBufferUpdate {
     template <typename T = u64>
     T DstAddress() const {
         ASSERT(update_memory.Value() == 1);
-        return reinterpret_cast<T>(dst_address_lo.Value() | u64(dst_address_hi & 0xFFFF) << 32);
+        return reinterpret_cast<T>(
+            static_cast<uintptr_t>(dst_address_lo.Value() | (u64(dst_address_hi & 0xFFFF) << 32)));
     }
 
     template <typename T = u64>
     T SrcAddress() const {
         ASSERT(source_select.Value() == SourceSelect::SrcAddress);
-        return reinterpret_cast<T>(src_address_lo.Value() | u64(src_address_hi & 0xFFFF) << 32);
+        return reinterpret_cast<T>(
+            static_cast<uintptr_t>(src_address_lo.Value() | (u64(src_address_hi & 0xFFFF) << 32)));
     }
 };
 
@@ -732,7 +734,7 @@ struct PM4CmdWriteData {
 
     template <typename T>
     T Address() const {
-        return reinterpret_cast<T>(addr64);
+        return reinterpret_cast<T>(static_cast<uintptr_t>(addr64));
     }
 };
 
@@ -763,7 +765,7 @@ struct PM4CmdEventWriteEos {
 
     template <typename T = u32*>
     T Address() const {
-        return reinterpret_cast<T>(address_lo | u64(address_hi) << 32);
+        return reinterpret_cast<T>(static_cast<uintptr_t>(address_lo | (u64(address_hi) << 32)));
     }
 
     u32 DataDWord() const {
@@ -822,7 +824,7 @@ struct PM4DumpConstRam {
 
     template <typename T>
     T Address() const {
-        return reinterpret_cast<T>((u64(addr_hi) << 32u) | addr_lo);
+        return reinterpret_cast<T>(static_cast<uintptr_t>((u64(addr_hi) << 32u) | addr_lo));
     }
 
     [[nodiscard]] u32 Offset() const {
@@ -874,7 +876,7 @@ struct PM4CmdIndirectBuffer {
 
     template <typename T>
     T* Address() const {
-        return reinterpret_cast<T*>((u64(ibase_hi) << 32u) | ibase_lo);
+        return reinterpret_cast<T*>(static_cast<uintptr_t>((u64(ibase_hi) << 32u) | ibase_lo));
     }
 };
 
@@ -911,7 +913,7 @@ struct PM4CmdReleaseMem {
 
     template <typename T>
     T Address() const {
-        return reinterpret_cast<T>(address_lo | u64(address_hi) << 32);
+        return reinterpret_cast<T>(static_cast<uintptr_t>(address_lo | (u64(address_hi) << 32)));
     }
 
     u32 DataDWord() const {
@@ -983,7 +985,8 @@ struct PM4CmdSetBase {
     T Address() const {
         ASSERT(base_index == BaseIndex::DisplayListPatchTable ||
                base_index == BaseIndex::DrawIndexIndirPatchTable);
-        return reinterpret_cast<T>(address0 | (u64(address1 & 0xffff) << 32u));
+        return reinterpret_cast<T>(
+            static_cast<uintptr_t>(address0 | (u64(address1 & 0xffff) << 32u)));
     }
 };
